refactor example into an integration test for ssl
